Update README.md
Problem Description

Break is right around the corner, and you’ve been planning a vacation for a while. You decide that you’re

going to travel via train, and you quickly realize that a train behaves similarly to a linked list! Using what

you’ve learned about the List ADT, you decide to implement it as a Train.

Solution Description

We highly recommend that you read this document in its entirety and familiarize yourself with the

requirements of this assignment before writing any code!

Important: All classes created in this assignment should be generic classes. All instances of generic

classes should specify a concrete type and should never use raw types. If a method returns a reference

type, it must use generic types, and should never use raw types.

A train is an ordered sequence of train cars. Each individual train car stores cargo and a reference to the

next train car in the sequence. The first train car is called the engine and a reference to it should be kept.

When there are no train cars attached to the train, the engine should be null